ecomapi/requirements.txt
graphene-django
django-graphiql
swagger_to
pyjwt